+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/zombiezen/graphql-server/compare/v0.4.0...HEAD
 
+### Changed
+
+-  `NewServer` now accepts functions as arguments which will be called to
+   resolve the top-level query or mutation objects. This can be used for
+   enforcing query cost restrictions or creating per-request objects. ([#39][])
+
+[#39]: https://github.com/zombiezen/graphql-server/issues/39
+
 ## [0.4.0][]
 
 The 0.4.0 makes usability improvements for field method implementations.